app/backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
django-environ

redis
celery
httpx
hypothesis
polyline

pytest
pytest-django
pytest-mock
pytest-timeout

# Stubs
django-stubs
djangorestframework-stubs
mypy

# auth
argon2-cffi
djangorestframework-simplejwt